backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
scikit-learn
google-generativeai
python-dotenv